Interface Setup
Two buttons to increase the player's experience in exchange for gold and to roll the player's hand, the chance of obtaining rare cards depends on the player's level.
Possibility of adding hand cards to the board.
If a card dies, it ends up in the graveyard.
Added a slider to display XP in real time.